requirements.txt
# YouTube Reporter v2.0 - 리팩토링된 의존성

# ===== 핵심 LangChain & LangGraph =====
langchain>=0.2.0,<0.3.0
langgraph>=0.1.1,<0.2.0
langchain-aws>=0.1.15
langchain-core>=0.2.2,<0.3.0
langchain-community>=0.2.0

# LangSmith (추적 및 모니터링)
langsmith>=0.1.0

# ===== 웹 프레임워크 =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ===== AWS 서비스 =====
boto3>=1.34.0
botocore>=1.34.0

# ===== HTTP 클라이언트 =====
requests>=2.31.0
httpx>=0.25.0

# ===== 설정 관리 =====
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ===== 타입 지원 =====
typing-extensions>=4.8.0

# ===== 비동기 지원 =====
aiofiles>=23.2.0

# ===== JSON 처리 =====
ujson>=5.8.0

# ===== 날짜/시간 =====
python-dateutil>=2.8.2

# ===== 로깅 및 디버깅 =====
colorlog>=6.8.0
rich>=13.7.0

# ===== 개발 도구 (선택사항) =====
# 코드 포맷팅
black>=23.11.0
isort>=5.12.0

# 코드 검사
flake8>=6.1.0
mypy>=1.7.0

# 테스트
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # 테스트용 HTTP 클라이언트

# ===== 선택적 의존성 =====
# OpenAI (DALL-E 등)
openai>=1.3.0

# 이미지 처리
Pillow>=10.1.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0

# 그래프 및 시각화
matplotlib>=3.8.0
seaborn>=0.13.0

# 웹 스크래핑 (선택적)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 캐싱 (선택적)
redis>=5.0.0

# 데이터베이스 (선택적)
sqlalchemy>=2.0.0
alembic>=1.13.0

# 보안
cryptography>=41.0.0

# 모니터링
prometheus-client>=0.19.0